Add support (and tests) for new affix called `dir:`
This new affix allows the `node-temp` user to select (on a per request basis)
another location (and hence file-system) where to create temporary files.

Although Mac & Windows users are mostly not impacted by this kind of change
(because TMP is usually located on the same volume as the main storage),
Linux users generally have a `/tmp` that is located on a separated file-system.
Moving to/from a temporary file-system leads to an EXDEV error in many cases.